[REFACTOR] Creates auth controller
